+++ packages/lesswrong/components/recentDiscussion/RecentDiscussionMeetupsPoke.tsx
@@ -8,7 +8,7 @@ import Button from '@material-ui/core/Button';
 import Checkbox from '@material-ui/core/Checkbox';
 import Input from '@material-ui/core/Input';
 import Geosuggest from 'react-geosuggest';
-import { pickBestReverseGeocodingResult } from '../../server/mapsUtils';
+import { pickBestReverseGeocodingResult } from '../../lib/geocoding';
 
 const styles = (theme: ThemeType): JssStyles => ({
   root: {
